# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Veridian Lead Management API
  version: 1.0.0
  description: Private consultancy enquiry-to-onboarding API. Better Auth owns routes under /api/auth.
servers:
  - url: http://localhost:5000
paths:
...
  /api/v1/admin/privacy-exports/{id}:
    get: { summary: Owner downloads one-time expiring subject export, security: [ { sessionCookie: [] } ], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { description: JSON export }, '404': { $ref: '#/components/responses/Error' } } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: veridian.session_token }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            fieldErrors: { type: object, additionalProperties: { type: array, items: { type: string } } }
  responses:
    Error: { description: Safe error envelope, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }